Move logfile setup into a separate file
